WIP: MAPISTATUS: Use MAPI_SUCCESS for success code name
It is hard now to search into log files for MAPI errors,
as we always put the prevfix MAPI_E_ even in case of success.

MS defines success for mapi ops to be SUCCESS_SUCCESS
and mark failures with MAPI_E_ (mapi error).
IMO, in our case MAPI_SUCCESS is a bit better.

Signed-off-by: Kamen Mazdrashki <kmazdrashki@zentyal.com>