emsmdbp/provision: Handle errors during Freebusy provisioning
This patch doesn't remove 'abort()' call in case of failure,
but instead print more information about the failure.
I have seen failures with this in cases when GlobalCount counter
gets messed up for some reason. In such cases, we better
abort as the state is messed up pretty badly and DB should be
fixed manually